+++ sertit/rasters_rio.py
@@ -854,8 +854,6 @@ def collocate(
     Collocate two georeferenced arrays:
     forces the *slave* raster to be exactly georeferenced onto the *master* raster by reprojection.
 
-    Use it like :code:`OTB SuperImpose`.
-
     .. code-block:: python
 
         >>> master_path = "path/to/master.tif"